+++ elements/share/src/types/index.ts
@@ -39,6 +39,8 @@ export type {
   IntegrationContentProps,
 } from './ShareEntities';
 
+export { TabType } from './ShareEntities';
+
 export type KeysOfType<T, TProp> = {
   [P in keyof T]: T[P] extends TProp ? P : never;
 }[keyof T];